console interaction implemented Hand filter returns a hand some fixes

+++ src/cards/CardCollection.scala
@@ -8,9 +8,7 @@ package cards
  * @author Oliver Friedrich
  */
 class CardCollection(pVector:Long) {
-  
-  
-  
+
   /**
    * The vector is the representation of the Cards. The 32 bits stand for 32 Cards and are set to 1 when the Collections contains the nth Card.
    * Initially the vector is 0 so no card is in the Collection
@@ -20,7 +18,7 @@ class CardCollection(pVector:Long) {
    * Adds a (or more) Card(s) to the Collection. If you want to add more than one Card with one call you have to OR all the Cards which shall be added.
    * @param pCard the bit vector representation of the Card(s) to add - 32 bits as Long
    */
-  def add(pCard: Long) = {
+  def add(pCard: Int) = {
     vector |= 1 << pCard
   }
   def addCards(pCardVector: Long) = {
@@ -30,7 +28,7 @@ class CardCollection(pVector:Long) {
    * Removes a (or more) Card(s) from the Collection. If you want to remove more than one Card with one call you have to OR all the Cards which shall be removed.
    * @param pCard the bit vector representation of the Card(s) to remove - 32 bits as Long
    */
-  def remove(pCard: Long) = {
+  def remove(pCard: Int) = {
     vector &= 1 << ~pCard
   }
   /**
@@ -38,7 +36,7 @@ class CardCollection(pVector:Long) {
    * @param pCard the bit vector representation of the Card(s) to test - 32 bits as Long
    * @return true if and only if the given Card(s) is/are in the Collection
    */
-  def contains(pCard: Long): Boolean = {
+  def contains(pCard: Int): Boolean = {
     (vector & 1 << pCard) != 0
   }
   
@@ -51,7 +49,7 @@ class CardCollection(pVector:Long) {
    * Gives the number of Cards in the Collection.
    * @return the number of Cards in the Collection
    */
-  def size: Long = {
+  def size: Int = {
     var x = vector
     var c = 0
     while (x != 0) {
@@ -69,10 +67,6 @@ class CardCollection(pVector:Long) {
     vector = 0
   }
   
-  
-  
-  
-
   /**
    * Standard empty constructor which initializes the Collection with 0 so empty.
    */
